version of write using _getBuffer
The code compiles and generate the file `dummy_bug.h5`. All the data get's written to the file except that we have to flatten the two dimensional array. At the moment if compile using the instructions as before and write h5dump then you should see something like this: 

HDF5 "bug_dummy.h5" {
GROUP "/" {
   DATASET "EventInfo_data_1" {
      DATATYPE  H5T_COMPOUND {
         H5T_STD_I32LE "Number";
      }
      DATASPACE  SIMPLE { ( 2 ) / ( 2 ) }
      DATA {
      (0): {
            20
         }, {
            21
         }
      }
   }
   DATASET "EventInfo_data_2" {
      DATATYPE  H5T_COMPOUND {
         H5T_STD_I32LE "Number";
      }
      DATASPACE  SIMPLE { ( 4 ) / ( 4 ) }
      DATA {
      (0): {
            22
         }, {
            23
         },
      (2): {
            24
         }, {
            25
         }
      }
   }
}
}

As you can see that the data is getting clumped together. An easy fix to this could be declaring a one dimensional array and then copying the relevant data and then giving it to HDF5 write method. If you would like me to try this, let me know.