writes ExampleMC along with EventInfoData
Also demonstrated the idea of creating new groups in the HDF5 file. This is how we will handle multiple event stores. For each event store we create /store_i (where i is some index = 1, 2, ...). And then the respective collection in that Event Store will be stored in the data set called /store_i/<name of collection datatype> ( for instance /store_1/EventInfoData ).