Introduce jinja2 template engine to generate code
Allows to remove some complexity from the generator by using a more
capable template engine that has the necessary logic built in.

For the moment a new jinja2 class generator is introduced and the
templates are placed into a separate directory. Hence, the two
generators briefly coexist here, but the jinja2 generator does not
necessarily yield c++ code that already compiles.

Introduce jinja2 templates and corresponding python code for
- Components
- Data
- Objects and ConstObjects
- Obj
- Collections
- selection .xml file