Make CollectionIDTable use a std::mutex
Previously used a std::recursive_mutex which is unnecessary in this
case, since the public member functions do not call each other, so there
should not be a way where the same thread tries to lock the same mutex a
second time.

Additionally made whatever can be made const const and now use
std::distance to calculate the index.