Make GenericParamters a concrete class
Since we do not currently inherit from it nor are there any plans to do
so in the future, there is no need to have virtual functions. This also
allows us to delete the virtual deconstructor and the compiler will now
generate all the constructors for us (rule of zero).