First attempt of splitting core and IO libraries
Making the core podio library independent of SIO (as it has previously
also been done with ROOT). In a similar fashion split off the SIO
dependent parts from the libTestDataModel and try to generate a specific
blocks library that only contains the SIOBlocks stuff. Currently need to
inclue at least one of the SIOBlock headers of the library to actually
force the proper self-registration of the SIOBlocks in the SIOBlock
factory.