Only override __getitem__ when necessary
For some reason cppyy seems to be calling the __getitem__ method with
only one argument instead of the usual two. Starting from root 6.22 this
breaks our python access to the collections. By only overriding it if it
is not already defined we can get it to work with root <= 6.20 and root
> 6.22.

Also removed the __iter__ overload, because that is not necessary to
loop over a collection, specifically because the iterator protocol would
also require a __next__ (or next in python2) method for actually
advancing.