Adjust the IReader interface and make implementations follow
Marking the getEntries and endOfEvent methods virtual and adding them to
the IReader interface allows to re-use the same code for the read tests
independent of the reader that actually provides the data.

NOTE: The current implementation of SIOReader::getEntries is just here
for feature completeness, and should be replaced with a more performant
and const-correct version.