Fix compiler warnings (#171)
* Fix pointer data member but no copy c'tor or assignment operator warnings

Wherever possible explicitly delete the two

* Fix all core Weffc++ warnings about members initialization list

This fixes all the warning: 'XYZ' should be initialized in the member
initialization list of the core classes.

It does not yet address the same warnings in the generated classes

* Fix Weffc++ warnings about member initialization for generated collections

* Fix Wshadow in tests as well as in generated collections

* Fix Wreorder by ordering the initialization list properly

* Fix Wunused warnings in core classes

* Fix Wunused warnings in generated code

* Fix Wnon-virtual-dtor and Wcatch-value warnings

* Turn typedefs into using statements remove unused code

* Fix Wsign-compares in core components and tests

* Use size_t and cast to int where necessary in generated collections

* Fix Wshadow warnings in core and generated classes

Lifted from #158

* Disable Weffc++ for collection constructors

Disbling it this way guarantees that the existing copy-constructor does
not get flagged, while other things, e.g. uninitialized members, still
are.

* Default initialize members to fix Weffc++

Since default initializing makes a class non-trivial, it is also no
longer a PDO. std::is_pod is deprecated in c++17 in any case, so replace
it with the properties we are actually interested in here.

* Make parser tests slightly more comprehensive

* Require -Werror in CI

Make it harder to accidentally merge new warnings from now on.

* Fix warnings from llvm

* Reduce selection.xml to minimum required

- Allows to delete the copy constructor of collections.
- Gets rid of warnings in the root generated Dict.cxx

* Remove guards for ignoring Weffc++ for generated collections

Now that we can delete the copy-constructor we are fine

* Make writers non-copyable

* Add Wno-overlength-strings for root generated dict

* Fix Wshadow warning for older gcc versions

gcc7 and gcc8 complain about shadowing a class member (function) with
the local data variable that is used in prepareAfterWrite. Renaming the
member to avoid having a naming clash, even though this is essentially a
false positive.

* Fix unused variable warnings in structured bindings

gcc7 and gcc8 on Ubuntu complained about these, maybe these are easier
to handle in a future c++ standard. For now, resort to using the pair
and .first and .second