Add templated access functionality to GenericParameters
Makes it easier to add new supported data types if ever necessary.
Additionally, gives this a more generic API that should make future
usage in the Frame easier.

These changes include a bunch of type helpers which now live in their
own header, which could also be useful in the future (or until we make
real use of concepts)