Use DL_PATHS again for correct test discovery env
Not sure why but this has become necessary again for running on Ubuntu24
in an LCG/dev3 environment to run test discovery succesfully. Otherwise
the environment for discovery is not set properly and running the
executable (during the build stage) fails with a symbol lookup error.
The latter is caused by the fact that, LCGs podio is not built with SIO
support.