Remove ReadOptions as dedicated struct
There is currently no *reasonable* need to have the possibility to steer
the behavior away from simply skipping unreadable collections. This is
technically a breaking change, but with corresponding changes to the
Frame we can almost restore the previous behavior (i.e. an exception on
an collection unavailable on fiel) and still fix bad crashes when we
truly have unreadable files / collections.