Skip unreadable collections instead of either crashing or skipping entire frames (#863)
* Make test read all collections

Remove the need to specifically state which collections to read

* Make the ROOTReader skip unreadable collections

* Make the RNTupleReader skip unreadable collections

* Add ReadOptions and incorporate collsToRead

Makes it possible to configure at runtime what should happen with
collections that can't be read, rather than making it a compile time
decision. Keeping old overloads undeprecated for the moment.

* Add some useful static constructors for good defaults

* Add ReadOptions to SIOReader

* Make Reader interface understand ReadOptions

* Make podio-dump be permissive by default

* Fix typo in member name

* Ensure header is installed

* Expose ReadOptions to python bindings

* Skip types that have changed enough to become unreadable

* Add docstrings for the new overloads and the options

* Skip one python unittest for old ROOT versions

* Harmonize warnings between TTree and RNTuple reading

* Make it possible to check non supported evolution behavior

* Adapt ignore list for sanitizer checks

* Remove ReadOptions as dedicated struct

There is currently no *reasonable* need to have the possibility to steer
the behavior away from simply skipping unreadable collections. This is
technically a breaking change, but with corresponding changes to the
Frame we can almost restore the previous behavior (i.e. an exception on
an collection unavailable on fiel) and still fix bad crashes when we
truly have unreadable files / collections.