Do not read ExtraCode for old schemas
Not necessary for the comparisons that we want to do as that doesn't
affect the Data PODs that we generate. Additionally, without this it is
required that the old schema files live at the same place as the current
schema which makes it impossible to store them in a subfolder.