Make the python writers have the same behavior as the C++ writers
If no collections are passed then let the C++ decide instead of
passing all collections (which is what the C++ writers do). In practice,
this only allows to write empty frames from python, currently not possible
as collections will evaluate to false and all the collections from
`frame.getAvailableCollections` will be passed.